vendor: Import v5.15.2.
Source URL:    git://code.qt.io/qt/qtwebengine-chromium.git
Source Commit: 8ca49e7ff5837b707c896157225f3e7f793d4943

+++ chromium/testing/test_env.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 # Copyright (c) 2012 The Chromium Authors. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
@@ -85,6 +85,9 @@ def get_sanitizer_env(cmd, asan, lsan, msan, tsan, cfi_diag):
     extra_env['LLVM_SYMBOLIZER_PATH'] = symbolizer_path
   else:
     symbolization_options = []
+
+  # Leverage sanitizer to print stack trace on abort (e.g. assertion failure).
+  symbolization_options.append('handle_abort=1')
 
   if asan:
     asan_options = symbolization_options[:]
@@ -174,22 +177,86 @@ def symbolize_snippets_in_json(cmd, env):
 
 
 def run_command_with_output(argv, stdoutfile, env=None, cwd=None):
-  """ Run command and stream its stdout/stderr to the console & |stdoutfile|.
+  """Run command and stream its stdout/stderr to the console & |stdoutfile|.
+
+  Also forward_signals to obey
+  https://chromium.googlesource.com/infra/luci/luci-py/+/master/appengine/swarming/doc/Bot.md#graceful-termination_aka-the-sigterm-and-sigkill-dance
+
+  Returns:
+    integer returncode of the subprocess.
   """
   print('Running %r in %r (env: %r)' % (argv, cwd, env))
   assert stdoutfile
   with io.open(stdoutfile, 'wb') as writer, \
       io.open(stdoutfile, 'rb', 1) as reader:
-    process = subprocess.Popen(argv, env=env, cwd=cwd, stdout=writer,
-        stderr=subprocess.STDOUT)
+    process = _popen(argv, env=env, cwd=cwd, stdout=writer,
+                     stderr=subprocess.STDOUT)
     forward_signals([process])
     while process.poll() is None:
       sys.stdout.write(reader.read())
+      # This sleep is needed for signal propagation. See the
+      # wait_with_signals() docstring.
       time.sleep(0.1)
     # Read the remaining.
     sys.stdout.write(reader.read())
     print('Command %r returned exit code %d' % (argv, process.returncode))
     return process.returncode
+
+
+def run_command(argv, env=None, cwd=None, log=True):
+  """Run command and stream its stdout/stderr both to stdout.
+
+  Also forward_signals to obey
+  https://chromium.googlesource.com/infra/luci/luci-py/+/master/appengine/swarming/doc/Bot.md#graceful-termination_aka-the-sigterm-and-sigkill-dance
+
+  Returns:
+    integer returncode of the subprocess.
+  """
+  if log:
+    print('Running %r in %r (env: %r)' % (argv, cwd, env))
+  process = _popen(argv, env=env, cwd=cwd, stderr=subprocess.STDOUT)
+  forward_signals([process])
+  return wait_with_signals(process)
+
+
+def run_command_output_to_handle(argv, file_handle, env=None, cwd=None):
+  """Run command and stream its stdout/stderr both to |file_handle|.
+
+  Also forward_signals to obey
+  https://chromium.googlesource.com/infra/luci/luci-py/+/master/appengine/swarming/doc/Bot.md#graceful-termination_aka-the-sigterm-and-sigkill-dance
+
+  Returns:
+    integer returncode of the subprocess.
+  """
+  print('Running %r in %r (env: %r)' % (argv, cwd, env))
+  process = _popen(
+      argv, env=env, cwd=cwd, stderr=file_handle, stdout=file_handle)
+  forward_signals([process])
+  exit_code = wait_with_signals(process)
+  print('Command returned exit code %d' % exit_code)
+  return exit_code
+
+
+def wait_with_signals(process):
+  """A version of process.wait() that works cross-platform.
+
+  This version properly surfaces the SIGBREAK signal.
+
+  From reading the subprocess.py source code, it seems we need to explicitly
+  call time.sleep(). The reason is that subprocess.Popen.wait() on Windows
+  directly calls WaitForSingleObject(), but only time.sleep() properly surface
+  the SIGBREAK signal.
+
+  Refs:
+  https://github.com/python/cpython/blob/v2.7.15/Lib/subprocess.py#L692
+  https://github.com/python/cpython/blob/v2.7.15/Modules/timemodule.c#L1084
+
+  Returns:
+    returncode of the process.
+  """
+  while process.poll() is None:
+    time.sleep(0.1)
+  return process.returncode
 
 
 def forward_signals(procs):
@@ -204,6 +271,8 @@ def forward_signals(procs):
   assert all(isinstance(p, subprocess.Popen) for p in procs)
   def _sig_handler(sig, _):
     for p in procs:
+      if p.poll() is not None:
+        continue
       # SIGBREAK is defined only for win32.
       if sys.platform == 'win32' and sig == signal.SIGBREAK:
         p.send_signal(signal.CTRL_BREAK_EVENT)
@@ -213,6 +282,7 @@ def forward_signals(procs):
     signal.signal(signal.SIGBREAK, _sig_handler)
   else:
     signal.signal(signal.SIGTERM, _sig_handler)
+    signal.signal(signal.SIGINT, _sig_handler)
 
 
 def run_executable(cmd, env, stdoutfile=None):
@@ -248,9 +318,9 @@ def run_executable(cmd, env, stdoutfile=None):
     # Symbolization works in-process on Windows even when sandboxed.
     use_symbolization_script = False
   else:
-    # LSan doesn't support sandboxing yet, so we use the in-process symbolizer.
-    # Note that ASan and MSan can work together with LSan.
-    use_symbolization_script = (asan or msan or cfi_diag) and not lsan
+    # If any sanitizer is enabled, we print unsymbolized stack trace
+    # that is required to run through symbolization script.
+    use_symbolization_script = (asan or msan or cfi_diag or lsan or tsan)
 
   if asan or lsan or msan or tsan or cfi_diag:
     extra_env.update(get_sanitizer_env(cmd, asan, lsan, msan, tsan, cfi_diag))
@@ -265,10 +335,17 @@ def run_executable(cmd, env, stdoutfile=None):
   cmd[0] = cmd[0].replace('/', os.path.sep)
   cmd = fix_python_path(cmd)
 
+  # We also want to print the GTEST env vars that were set by the caller,
+  # because you need them to reproduce the task properly.
+  env_to_print = extra_env.copy()
+  for env_var_name in ('GTEST_SHARD_INDEX', 'GTEST_TOTAL_SHARDS'):
+      if env_var_name in env:
+          env_to_print[env_var_name] = env[env_var_name]
+
   print('Additional test environment:\n%s\n'
         'Command: %s\n' % (
         '\n'.join('    %s=%s' %
-            (k, v) for k, v in sorted(extra_env.iteritems())),
+            (k, v) for k, v in sorted(env_to_print.iteritems())),
         ' '.join(cmd)))
   sys.stdout.flush()
   env.update(extra_env or {})
@@ -279,25 +356,31 @@ def run_executable(cmd, env, stdoutfile=None):
     elif use_symbolization_script:
       # See above comment regarding offline symbolization.
       # Need to pipe to the symbolizer script.
-      p1 = subprocess.Popen(cmd, env=env, stdout=subprocess.PIPE,
-                            stderr=sys.stdout)
-      p2 = subprocess.Popen(
+      p1 = _popen(cmd, env=env, stdout=subprocess.PIPE,
+                  stderr=sys.stdout)
+      p2 = _popen(
           get_sanitizer_symbolize_command(executable_path=cmd[0]),
           env=env, stdin=p1.stdout)
       p1.stdout.close()  # Allow p1 to receive a SIGPIPE if p2 exits.
       forward_signals([p1, p2])
-      p1.wait()
-      p2.wait()
+      wait_with_signals(p1)
+      wait_with_signals(p2)
       # Also feed the out-of-band JSON output to the symbolizer script.
       symbolize_snippets_in_json(cmd, env)
       return p1.returncode
     else:
-      p = subprocess.Popen(cmd, env=env)
-      forward_signals([p])
-      return p.wait()
+      return run_command(cmd, env=env, log=False)
   except OSError:
     print >> sys.stderr, 'Failed to start %s' % cmd
     raise
+
+
+def _popen(*args, **kwargs):
+  assert 'creationflags' not in kwargs
+  if sys.platform == 'win32':
+    # Necessary for signal handling. See crbug.com/733612#c6.
+    kwargs['creationflags'] = subprocess.CREATE_NEW_PROCESS_GROUP
+  return subprocess.Popen(*args, **kwargs)
 
 
 def main():
